refactor(site): share currency helpers
